# Conflicts:
#	api/v1/testkube.yaml
#	cmd/api-server/main.go
#	cmd/kubectl-testkube/commands/testsuites/run.go
#	docs/docs/articles/crds.md
#	docs/docs/articles/creating-test-suites.md
#	docs/docs/articles/getting-test-suites-results.md
#	go.mod
#	go.sum
#	internal/app/api/v1/testsuites.go
#	pkg/api/v1/client/interface.go
#	pkg/api/v1/client/testsuite.go
#	pkg/api/v1/testkube/model_execution_extended.go
#	pkg/api/v1/testkube/model_test_suite_execution_request.go
#	pkg/api/v1/testkube/model_test_suite_execution_update_request.go
#	pkg/triggers/watcher.go
#	test/suites/demo-testsuite.json
#	test/suites/executor-cypress-smoke-tests.json
#	test/suites/executor-jmeter-smoke-tests.json
#	test/suites/executor-k6-other-tests.json
#	test/suites/executor-k6-smoke-tests.json
#	test/suites/executor-postman-smoke-tests.json
#	test/suites/staging-testsuite.json

+++ internal/app/api/v1/tests.go
@@ -8,39 +8,45 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/kubeshop/testkube/pkg/repository/result"
-
 	"github.com/gofiber/fiber/v2"
 	"go.mongodb.org/mongo-driver/mongo"
+	"k8s.io/apimachinery/pkg/api/errors"
 
 	testsv3 "github.com/kubeshop/testkube-operator/apis/tests/v3"
 	"github.com/kubeshop/testkube-operator/client/tests/v3"
+	"github.com/kubeshop/testkube-operator/pkg/secret"
 	"github.com/kubeshop/testkube/pkg/api/v1/testkube"
 	"github.com/kubeshop/testkube/pkg/crd"
 	"github.com/kubeshop/testkube/pkg/executor/client"
 	executionsmapper "github.com/kubeshop/testkube/pkg/mapper/executions"
 	testsmapper "github.com/kubeshop/testkube/pkg/mapper/tests"
-
-	"k8s.io/apimachinery/pkg/api/errors"
+	"github.com/kubeshop/testkube/pkg/repository/result"
 )
 
 // GetTestHandler is method for getting an existing test
 func (s TestkubeAPI) GetTestHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
 		name := c.Params("id")
+		if name == "" {
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("failed to get test: id cannot be empty"))
+		}
+		errPrefix := fmt.Sprintf("failed to get test %s", name)
 		crTest, err := s.TestsClient.Get(name)
 		if err != nil {
 			if errors.IsNotFound(err) {
-				return s.Error(c, http.StatusNotFound, err)
+				return s.Error(c, http.StatusNotFound, fmt.Errorf("%s: client was unable to find test: %w", errPrefix, err))
 			}
 
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: client failed to find test: %w", errPrefix, err))
 		}
 
 		test := testsmapper.MapTestCRToAPI(*crTest)
 		if c.Accepts(mediaTypeJSON, mediaTypeYAML) == mediaTypeYAML {
 			test.QuoteTestTextFields()
 			data, err := crd.GenerateYAML(crd.TemplateTest, []testkube.Test{test})
+			if err != nil {
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not build CRD: %w", errPrefix, err))
+			}
 			return s.getCRDs(c, data, err)
 		}
 
@@ -52,31 +58,38 @@ func (s TestkubeAPI) GetTestHandler() fiber.Handler {
 func (s TestkubeAPI) GetTestWithExecutionHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
 		name := c.Params("id")
+		if name == "" {
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("failed to get test with execution: id cannot be empty"))
+		}
+		errPrefix := fmt.Sprintf("failed to get test %s with execution", name)
 		crTest, err := s.TestsClient.Get(name)
 		if err != nil {
 			if errors.IsNotFound(err) {
-				return s.Error(c, http.StatusNotFound, err)
+				return s.Error(c, http.StatusNotFound, fmt.Errorf("%s: client failed to find test: %w", errPrefix, err))
 			}
 
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: client failed to find test: %w", errPrefix, err))
 		}
 
 		test := testsmapper.MapTestCRToAPI(*crTest)
 		if c.Accepts(mediaTypeJSON, mediaTypeYAML) == mediaTypeYAML {
 			test.QuoteTestTextFields()
 			data, err := crd.GenerateYAML(crd.TemplateTest, []testkube.Test{test})
+			if err != nil {
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not build CRD: %w", errPrefix, err))
+			}
 			return s.getCRDs(c, data, err)
 		}
 
 		ctx := c.Context()
 		startExecution, startErr := s.ExecutionResults.GetLatestByTest(ctx, name, "starttime")
 		if startErr != nil && startErr != mongo.ErrNoDocuments {
-			return s.Error(c, http.StatusInternalServerError, startErr)
+			return s.Error(c, http.StatusInternalServerError, fmt.Errorf("%s: failed to get executions by start time: %w", errPrefix, startErr))
 		}
 
 		endExecution, endErr := s.ExecutionResults.GetLatestByTest(ctx, name, "endtime")
 		if endErr != nil && endErr != mongo.ErrNoDocuments {
-			return s.Error(c, http.StatusInternalServerError, endErr)
+			return s.Error(c, http.StatusInternalServerError, fmt.Errorf("%s: failed to get executions by end time: %w", errPrefix, endErr))
 		}
 
 		testWithExecution := testkube.TestWithExecution{
@@ -102,7 +115,7 @@ func (s TestkubeAPI) getFilteredTestList(c *fiber.Ctx) (*testsv3.TestList, error
 
 	crTests, err := s.TestsClient.List(c.Query("selector"))
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("client failed to list tests: %w", err)
 	}
 
 	search := c.Query("textSearch")
@@ -131,9 +144,10 @@ func (s TestkubeAPI) getFilteredTestList(c *fiber.Ctx) (*testsv3.TestList, error
 // ListTestsHandler is a method for getting list of all available tests
 func (s TestkubeAPI) ListTestsHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		errPrefix := "failed to list tests"
 		crTests, err := s.getFilteredTestList(c)
 		if err != nil {
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: unable to get filtered tests: %w", errPrefix, err))
 		}
 
 		tests := testsmapper.MapTestListKubeToAPI(*crTests)
@@ -143,6 +157,9 @@ func (s TestkubeAPI) ListTestsHandler() fiber.Handler {
 			}
 
 			data, err := crd.GenerateYAML(crd.TemplateTest, tests)
+			if err != nil {
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not build CRD: %w", errPrefix, err))
+			}
 			return s.getCRDs(c, data, err)
 		}
 
@@ -169,7 +186,7 @@ func (s TestkubeAPI) TestMetricsHandler() fiber.Handler {
 
 		metrics, err := s.ExecutionResults.GetTestMetrics(context.Background(), testName, limit, last)
 		if err != nil {
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusInternalServerError, fmt.Errorf("failed to get metrics for test %s: %w", testName, err))
 		}
 
 		return c.JSON(metrics)
@@ -180,7 +197,7 @@ func (s TestkubeAPI) TestMetricsHandler() fiber.Handler {
 func (s TestkubeAPI) getLatestExecutions(ctx context.Context, testNames []string) (map[string]testkube.Execution, error) {
 	executions, err := s.ExecutionResults.GetLatestByTests(ctx, testNames, "starttime")
 	if err != nil && err != mongo.ErrNoDocuments {
-		return nil, err
+		return nil, fmt.Errorf("could not get latest executions for tests %s sorted by start time: %w", testNames, err)
 	}
 
 	startExecutionMap := make(map[string]testkube.Execution, len(executions))
@@ -190,7 +207,7 @@ func (s TestkubeAPI) getLatestExecutions(ctx context.Context, testNames []string
 
 	executions, err = s.ExecutionResults.GetLatestByTests(ctx, testNames, "endtime")
 	if err != nil && err != mongo.ErrNoDocuments {
-		return nil, err
+		return nil, fmt.Errorf("could not get latest executions for tests %s sorted by end time: %w", testNames, err)
 	}
 
 	endExecutionMap := make(map[string]testkube.Execution, len(executions))
@@ -229,9 +246,10 @@ func (s TestkubeAPI) getLatestExecutions(ctx context.Context, testNames []string
 // ListTestWithExecutionsHandler is a method for getting list of all available test with latest executions
 func (s TestkubeAPI) ListTestWithExecutionsHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		errPrefix := "failed to list tests with executions"
 		crTests, err := s.getFilteredTestList(c)
 		if err != nil {
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: unable to get filtered tests: %w", errPrefix, err))
 		}
 
 		tests := testsmapper.MapTestListKubeToAPI(*crTests)
@@ -241,6 +259,9 @@ func (s TestkubeAPI) ListTestWithExecutionsHandler() fiber.Handler {
 			}
 
 			data, err := crd.GenerateYAML(crd.TemplateTest, tests)
+			if err != nil {
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not build CRD: %w", errPrefix, err))
+			}
 			return s.getCRDs(c, data, err)
 		}
 
@@ -253,7 +274,7 @@ func (s TestkubeAPI) ListTestWithExecutionsHandler() fiber.Handler {
 
 		executionMap, err := s.getLatestExecutions(ctx, testNames)
 		if err != nil {
-			return s.Error(c, http.StatusInternalServerError, err)
+			return s.Error(c, http.StatusInternalServerError, fmt.Errorf("%s: could not get latest executions: %w", errPrefix, err))
 		}
 
 		for i := range tests {
@@ -293,7 +314,7 @@ func (s TestkubeAPI) ListTestWithExecutionsHandler() fiber.Handler {
 		if status != "" {
 			statusList, err := testkube.ParseExecutionStatusList(status, ",")
 			if err != nil {
-				return s.Error(c, http.StatusBadRequest, fmt.Errorf("execution status filter invalid: %w", err))
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: execution status filter invalid: %w", errPrefix, err))
 			}
 
 			statusMap := statusList.ToMap()
@@ -315,7 +336,7 @@ func (s TestkubeAPI) ListTestWithExecutionsHandler() fiber.Handler {
 			pageSize = result.PageDefaultLimit
 			page, err = strconv.Atoi(pageParam)
 			if err != nil {
-				return s.Error(c, http.StatusBadRequest, fmt.Errorf("test page filter invalid: %w", err))
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: test page filter invalid: %w", errPrefix, err))
 			}
 		}
 
@@ -323,7 +344,7 @@ func (s TestkubeAPI) ListTestWithExecutionsHandler() fiber.Handler {
 		if pageSizeParam != "" {
 			pageSize, err = strconv.Atoi(pageSizeParam)
 			if err != nil {
-				return s.Error(c, http.StatusBadRequest, fmt.Errorf("test page size filter invalid: %w", err))
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: test page size filter invalid: %w", errPrefix, err))
 			}
 		}
 
@@ -346,23 +367,32 @@ func (s TestkubeAPI) ListTestWithExecutionsHandler() fiber.Handler {
 // CreateTestHandler creates new test CR based on test content
 func (s TestkubeAPI) CreateTestHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		errPrefix := "failed to create test"
 
 		var request testkube.TestUpsertRequest
 		err := c.BodyParser(&request)
 		if err != nil {
-			return s.Error(c, http.StatusBadRequest, err)
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: failed to unmarshal request: %w", errPrefix, err))
+		}
+		err = testkube.ValidateUpsertTestRequest(request)
+		if err != nil {
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: invalid test: %w", errPrefix, err))
 		}
 
+		errPrefix = errPrefix + " " + request.Name
 		if request.ExecutionRequest != nil && request.ExecutionRequest.Args != nil {
 			request.ExecutionRequest.Args, err = testkube.PrepareExecutorArgs(request.ExecutionRequest.Args)
 			if err != nil {
-				return s.Error(c, http.StatusBadRequest, err)
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not prepare executor args: %w", errPrefix, err))
 			}
 		}
 
 		if c.Accepts(mediaTypeJSON, mediaTypeYAML) == mediaTypeYAML {
 			request.QuoteTestTextFields()
 			data, err := crd.GenerateYAML(crd.TemplateTest, []testkube.TestUpsertRequest{request})
+			if err != nil {
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not build CRD: %w", errPrefix, err))
+			}
 			return s.getCRDs(c, data, err)
 		}
 
@@ -372,7 +402,7 @@ func (s TestkubeAPI) CreateTestHandler() fiber.Handler {
 		test.Namespace = s.Namespace
 		var secrets map[string]string
 		if request.Content != nil && request.Content.Repository != nil {
-			secrets = getTestSecretsData(request.Content.Repository.Username, request.Content.Repository.Token)
+			secrets = createTestSecretsData(request.Content.Repository.Username, request.Content.Repository.Token)
 		}
 
 		createdTest, err := s.TestsClient.Create(test, tests.Option{Secrets: secrets})
@@ -380,7 +410,7 @@ func (s TestkubeAPI) CreateTestHandler() fiber.Handler {
 		s.Metrics.IncCreateTest(test.Spec.Type_, err)
 
 		if err != nil {
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: client could not create test: %w", errPrefix, err))
 		}
 
 		c.Status(http.StatusCreated)
@@ -391,31 +421,37 @@ func (s TestkubeAPI) CreateTestHandler() fiber.Handler {
 // UpdateTestHandler updates an existing test CR based on test content
 func (s TestkubeAPI) UpdateTestHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		errPrefix := "failed to update test"
 		var request testkube.TestUpdateRequest
 		err := c.BodyParser(&request)
 		if err != nil {
-			return s.Error(c, http.StatusBadRequest, err)
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not parse request: %w", errPrefix, err))
 		}
-
 		var name string
 		if request.Name != nil {
 			name = *request.Name
+			errPrefix = errPrefix + " " + name
+		}
+
+		err = testkube.ValidateUpdateTestRequest(request)
+		if err != nil {
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: invalid test: %w", errPrefix, err))
 		}
 
 		// we need to get resource first and load its metadata.ResourceVersion
 		test, err := s.TestsClient.Get(name)
 		if err != nil {
 			if errors.IsNotFound(err) {
-				return s.Error(c, http.StatusNotFound, err)
+				return s.Error(c, http.StatusNotFound, fmt.Errorf("%s: client could not find test: %w", errPrefix, err))
 			}
 
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: client could not get test: %w", errPrefix, err))
 		}
 
 		if request.ExecutionRequest != nil && *request.ExecutionRequest != nil && (*request.ExecutionRequest).Args != nil {
 			*(*request.ExecutionRequest).Args, err = testkube.PrepareExecutorArgs(*(*request.ExecutionRequest).Args)
 			if err != nil {
-				return s.Error(c, http.StatusBadRequest, err)
+				return s.Error(c, http.StatusBadRequest, fmt.Errorf("%s: could not prepare executor arguments: %w", errPrefix, err))
 			}
 		}
 
@@ -429,16 +465,12 @@ func (s TestkubeAPI) UpdateTestHandler() fiber.Handler {
 			username := (*(*request.Content).Repository).Username
 			token := (*(*request.Content).Repository).Token
 			if username != nil || token != nil {
-				var uValue, tValue string
-				if username != nil {
-					uValue = *username
+				data, err := s.SecretClient.Get(secret.GetMetadataName(name, client.SecretTest))
+				if err != nil && !errors.IsNotFound(err) {
+					return s.Error(c, http.StatusBadGateway, err)
 				}
 
-				if token != nil {
-					tValue = *token
-				}
-
-				option = &tests.Option{Secrets: getTestSecretsData(uValue, tValue)}
+				option = &tests.Option{Secrets: updateTestSecretsData(data, username, token)}
 			}
 		}
 
@@ -449,10 +481,10 @@ func (s TestkubeAPI) UpdateTestHandler() fiber.Handler {
 			updatedTest, err = s.TestsClient.Update(testSpec)
 		}
 
-		s.Metrics.IncUpdateTest(updatedTest.Spec.Type_, err)
+		s.Metrics.IncUpdateTest(testSpec.Spec.Type_, err)
 
 		if err != nil {
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: client could not update test %w", errPrefix, err))
 		}
 
 		return c.JSON(updatedTest)
@@ -463,37 +495,76 @@ func (s TestkubeAPI) UpdateTestHandler() fiber.Handler {
 func (s TestkubeAPI) DeleteTestHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
 		name := c.Params("id")
+		if name == "" {
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("failed to delete test: id cannot be empty"))
+		}
+		errPrefix := fmt.Sprintf("failed to delete test %s", name)
 		err := s.TestsClient.Delete(name)
 		if err != nil {
 			if errors.IsNotFound(err) {
-				return s.Warn(c, http.StatusNotFound, err)
+				return s.Warn(c, http.StatusNotFound, fmt.Errorf("%s: client could not find test: %w", errPrefix, err))
 			}
 
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: client could not delete test: %w", errPrefix, err))
 		}
 
 		// delete executions for test
 		if err = s.ExecutionResults.DeleteByTest(c.Context(), name); err != nil {
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusInternalServerError, fmt.Errorf("%s: could not delete the executions of the test: %w", errPrefix, err))
 		}
 
 		return c.SendStatus(http.StatusNoContent)
 	}
 }
 
+// AbortTestHandler is a method for aborting a executions of a test with id
+func (s TestkubeAPI) AbortTestHandler() fiber.Handler {
+	return func(c *fiber.Ctx) error {
+		ctx := c.Context()
+		name := c.Params("id")
+		if name == "" {
+			return s.Error(c, http.StatusBadRequest, fmt.Errorf("failed to abort test: id cannot be empty"))
+		}
+		errPrefix := fmt.Sprintf("failed to abort test %s", name)
+		filter := result.NewExecutionsFilter().WithTestName(name).WithStatus(string(testkube.RUNNING_ExecutionStatus))
+		executions, err := s.ExecutionResults.GetExecutions(ctx, filter)
+		if err != nil {
+			if err == mongo.ErrNoDocuments {
+				return s.Error(c, http.StatusNotFound, fmt.Errorf("%s: executions with name %s not found", errPrefix, name))
+			}
+			return s.Error(c, http.StatusInternalServerError, fmt.Errorf("%s: could not get executions: %w", errPrefix, err))
+		}
+
+		var results []testkube.ExecutionResult
+		for _, execution := range executions {
+			res, errAbort := s.Executor.Abort(ctx, &execution)
+			if errAbort != nil {
+				s.Log.Errorw("aborting execution failed", "execution", execution, "error", errAbort)
+				err = errAbort
+			}
+			s.Metrics.IncAbortTest(execution.TestType, res.IsFailed())
+			results = append(results, *res)
+		}
+
+		return c.JSON(results)
+	}
+}
+
 // DeleteTestsHandler for deleting all tests
 func (s TestkubeAPI) DeleteTestsHandler() fiber.Handler {
 	return func(c *fiber.Ctx) error {
+		errPrefix := "failed to delete tests"
 		var err error
 		var testNames []string
 		selector := c.Query("selector")
 		if selector == "" {
 			err = s.TestsClient.DeleteAll()
 		} else {
-			testList, err := s.TestsClient.List(selector)
+			var testList *testsv3.TestList
+			testList, err = s.TestsClient.List(selector)
 			if err != nil {
 				if !errors.IsNotFound(err) {
-					return s.Error(c, http.StatusBadGateway, err)
+					return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: client could not list tests: %w", errPrefix, err))
 				}
 			} else {
 				for _, item := range testList.Items {
@@ -506,10 +577,10 @@ func (s TestkubeAPI) DeleteTestsHandler() fiber.Handler {
 
 		if err != nil {
 			if errors.IsNotFound(err) {
-				return s.Warn(c, http.StatusNotFound, err)
+				return s.Warn(c, http.StatusNotFound, fmt.Errorf("%s: could not find tests: %w", errPrefix, err))
 			}
 
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusBadGateway, fmt.Errorf("%s: could not delete tests: %w", errPrefix, err))
 		}
 
 		// delete all executions for tests
@@ -520,14 +591,14 @@ func (s TestkubeAPI) DeleteTestsHandler() fiber.Handler {
 		}
 
 		if err != nil {
-			return s.Error(c, http.StatusBadGateway, err)
+			return s.Error(c, http.StatusInternalServerError, fmt.Errorf("%s: could not delete executions: %w", errPrefix, err))
 		}
 
 		return c.SendStatus(http.StatusNoContent)
 	}
 }
 
-func getTestSecretsData(username, token string) map[string]string {
+func createTestSecretsData(username, token string) map[string]string {
 	if username == "" && token == "" {
 		return nil
 	}
@@ -539,6 +610,30 @@ func getTestSecretsData(username, token string) map[string]string {
 
 	if token != "" {
 		data[client.GitTokenSecretName] = token
+	}
+
+	return data
+}
+
+func updateTestSecretsData(data map[string]string, username, token *string) map[string]string {
+	if data == nil {
+		data = make(map[string]string)
+	}
+
+	if username != nil {
+		if *username == "" {
+			delete(data, client.GitUsernameSecretName)
+		} else {
+			data[client.GitUsernameSecretName] = *username
+		}
+	}
+
+	if token != nil {
+		if *token == "" {
+			delete(data, client.GitTokenSecretName)
+		} else {
+			data[client.GitTokenSecretName] = *token
+		}
 	}
 
 	return data
